Add Psych 4 compatibility (Ruby 3.1)
Psych 4 changed the `load` method to call `safe_load`, which does not
natively permit BigDecimal instances, which are used in the
`derived_unit.yaml` file. Since the YAML files come from the gem itself,
it seems reasonable to trust them, so now we load them via
`unsafe_load`, which permits BigDecimal instances.

Psych 4 shipped with Ruby 3.1, however it's compatible with Ruby 2.4+.
`unsafe_load` was backported to Psych 3.3.2, however it does not exist
in prior versions. For now I'm rescuing a `NoMethodError` on calling
`unsafe_load` to then fall back to `load` for backwards compatibility.